backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
redis>=3.5.3,<5.0.0
httpx>=0.24.0,<0.25.0
supabase==2.0.3
pydantic>=1.10.2,<2.1.0
pydantic-settings>=1.2.0,<2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
langchain-core>=0.1.0
langchain-anthropic>=0.1.1
anthropic>=0.39.0,<1.0.0
redis-om==0.2.1
pillow==10.2.0
pypdf==4.0.1
python-magic==0.4.27 